requirements.txt
google-generativeai
mistralai
anthropic
